packages/mcp-server/pyproject.toml
[project]
name = "claude-vault-mcp"
version = "1.0.0"
description = "MCP server for HashiCorp Vault secret management via claude-vault"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.1.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "webauthn>=2.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
claude-vault-mcp = "claude_vault_mcp:run"
vault-approve-server = "claude_vault_mcp.approval_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_vault_mcp"]
